+++ packages/core/src/utils/utilities.ts
@@ -108,3 +108,49 @@ export const buildFormData = (
 
 export const toArray = <T>(x: T[] | T | undefined) =>
   Array.isArray(x) ? x : x ? [x] : []
+
+/**
+ * Formats the file name to a more user-friendly version.
+ * @param fileName - The original file name.
+ * @returns The formatted file name.
+ * @example
+ * formatFileName('smartphoneX5-23-256gb-black-edition-2023.xlsx')
+ * // Returns: 'smartphoneX5-23...ck-edition-2023.xlsx'
+ */
+export function formatFileName(fileName: string): string {
+  const maxLength = 30
+  const extensionIndex = fileName.lastIndexOf('.')
+  const extension = extensionIndex !== -1 ? fileName.slice(extensionIndex) : ''
+
+  if (fileName.length <= maxLength) {
+    return fileName
+  }
+
+  let nameWithoutExtension =
+    extensionIndex !== -1 ? fileName.slice(0, extensionIndex) : fileName
+
+  if (nameWithoutExtension.length > maxLength) {
+    const start = nameWithoutExtension.slice(0, maxLength / 2)
+    const end = nameWithoutExtension.slice(-maxLength / 2)
+    nameWithoutExtension = `${start}...${end}`
+  }
+
+  return `${nameWithoutExtension}${extension}`
+}
+
+/**
+ * Formats a file size in bytes to a human-readable string.
+ * @param bytes - The file size in bytes.
+ * @returns The formatted file size string.
+ * @example
+ * formatFileSize(2048) // Returns: '2 KB'
+ */
+export const formatFileSize = (bytes: number): string => {
+  if (bytes === 0) return '0 Bytes'
+
+  const k = 1024
+  const sizes = ['Bytes', 'KB', 'MB', 'GB']
+  const i = Math.floor(Math.log(bytes) / Math.log(k))
+
+  return Number.parseFloat((bytes / k ** i).toFixed(2)) + ' ' + sizes[i]
+}